% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_helper.R
\name{fair_generate_report}
\alias{fair_generate_report}
\title{Generate a report for large dataset using multiple slides}
\usage{
fair_generate_report(
  raw_data = NULL,
  chunk_size = 10,
  report_author = "Fair Analytics"
)
}
\arguments{
\item{raw_data}{the large dataset to use to generate the slides}

\item{chunk_size}{the regular size of subset in each slide}

\item{report_author}{The author of the report}
}
\description{
t.b.d
}
\author{
Fair Analytics
}
